+++ api/nutshell.json
@@ -142,7 +142,7 @@
   "newsArticles": [
     {
       "id": 1,
-      "timeStamp":"20/05/2020, 17:30:42",
+      "timeStamp": "21/05/2020, 15:50:07",
       "headline": "Club Omaha Plans to Reopen but with Masks and Gloves",
       "synopsis": "The dancers at Omaha Adult Entertainment Club are returning back to work next week, but they will be a little more covered than they normally are. The dancers will shed...",
       "url": "https://www.oddee.com/club-omaha-plans-to-reopen-but-with-masks-and-gloves-52701/",
@@ -150,18 +150,10 @@
     },
     {
       "id": 2,
-      "timeStamp":"20/05/2020, 17:30:42",
+      "timeStamp": "20/05/2020, 17:30:42",
       "headline": "Man In India Worships Trump As A God",
       "synopsis": "The Indian Constitution guarantees freedom of religion, but some people take it too far. Even to the extent of worshiping Trump... Bussa Krishna Before Donald Trump became a leader of...",
       "url": "https://www.oddee.com/man-in-india-worships-trump-as-a-god-51659/",
-      "userId": 2
-    },
-    {
-      "id": 4,
-      "timeStamp":"20/05/2020, 17:30:42",
-      "headline": "edit",
-      "synopsis": "edit",
-      "url": "edit",
       "userId": 2
     }
   ],
