+++ src/index.html
@@ -57,10 +57,7 @@
 
 
 
-<script src="news.js"></script>
-<script src="script.js"></script>
-
-<!-- <script type="module" src="script.js"></script> -->
+<script type="module" src="script.js"></script>
 </body>
 </html>
