+++ src/tasks-DOMPrinter.js
@@ -27,4 +27,4 @@ const printAllTasks = (tasksArray) => {
 
 
 
-// export default printAllTasks;
+export default printAllTasks;
